+++ comp2013/src/main/java/comp2013/View/IView.java
@@ -49,4 +49,15 @@ public interface IView {
      * Restarts the game by resetting the UI.
      */
     void restartGame();
+
+
+    /**
+     * Pause the game when ESC is pressed.
+     */
+    void pauseGame();
+
+    /**
+     * Handles the clearing of the pause UI when the game is unpaused.
+     */
+    void unpauseGame();
 }